Resolve htmlproofer failures in CI (#70)
* Update URLs to use https instead of https

This is a bulk update to resolve almost all of the htmlproofer CI errors
(and generally to link to secure versions of websites). The remaining 7
errors reported by htmlproofer are conference-specific and relate to the
inability to validate links via libcurl, which may be due to servers not
supporting https or domains no longer resolving.

* Update websites for a batch of conferences

These updates fall into two categories:

1. Websites of inactive conferences that no longer exist or no longer
   contain conference-related content. In these instances, I've updated
   the links to the latest relevant snapshots I could find on the
   Internet Archive.

2. Conferences whose canonical domains (as referenced by their websites
   or social media accounts) do not match the domains used here
   currently.

The last remaining htmlproofer error regards PyCon UK, whose canonical
domain is correctly listed here, but it does not respond to https
requests, and so I've reverted it back to http in this commit. This
ensures that the link is still usable but does not resolve the CI error
regarding http links.

* Link to 2023.pyconuk.org